refactor: Move to Chatroom Controller and Service
